add docstrings and use a glue plugin
