+++ static/js/lib/ckeditor/plugins/ResourceLinkMarkdownSyntax.ts
@@ -12,17 +12,33 @@ import { RESOURCE_LINK_CONFIG_KEY } from "./constants"
 
 const RESOURCE_LINK_SHORTCODE_REGEX = Shortcode.regex("resource_link", true)
 
+const queryKeys = {
+  uuid:   "ocw_resource_link_uuid",
+  suffix: "ocw_resource_link_suffix"
+}
+
 /**
- * Class for defining Markdown conversion rules for Resource links
+ * Class for defining Markdown conversion rules for Resource links.
+ *  - Converts from markdown to HTML when populating CKEditor
+ *  - Converts from HTML to markdown when saving
  *
- * These are stored in Markdown like this:
+ * Resource links are stored in markdown as:
  *
  * ```md
- * {{% resource_link AUUIDUNLIKEANYOTHER "Here's a link to my resource" %}}
+ * {{% resource_link "some-uuid" "Title of Link" "optional-link-suffix" %}}
  * ```
  *
- * The first argument is the uuid of the resource to which we're linking, and
- * the second argument is that text that should be rendered inside of the link.
+ * and in HTML as:
+ * ```html
+ * <a href="{previewURL} \
+ *  ?ocw_resource_link_uuid={some-uuid} \
+ *  &ocw_resource_link_suffix={optional-link-suffix}"
+ * > Title of Link </a>
+ * ```
+ * Where:
+ *  - {previewURL} portion is ONLY used for previewing the resource in CKEditor
+ *    and is generated by a template specified in CKEditor's `resource-link`
+ *    configuration item.
  *
  */
 export default class ResourceLinkMarkdownSyntax extends MarkdownSyntaxPlugin {
@@ -50,20 +66,21 @@ export default class ResourceLinkMarkdownSyntax extends MarkdownSyntaxPlugin {
 
   isResourceLinkHref = (href?: string): boolean => {
     if (!href) return false
-    return href.includes("?ocw_resource_link_uuid=")
+    const url = new URL(href)
+    return url.searchParams.has(queryKeys.uuid)
   }
 
-  makeResourceLinkHref = (uuid: string, fragment = "") => {
+  makeResourceLinkHref = (uuid: string, suffix = "") => {
     const href = new URL(this.hrefTemplate.replace(/:uuid/g, uuid))
-    href.searchParams.set("ocw_resource_link_uuid", uuid)
-    href.searchParams.set("ocw_resource_link_fragment", fragment)
+    href.searchParams.set(queryKeys.uuid, uuid)
+    href.searchParams.set(queryKeys.suffix, suffix)
     return href.toString()
   }
 
   removeResourceLinkQueryParams = (href: string): string => {
     const url = new URL(href)
-    url.searchParams.delete("ocw_resource_link_uuid")
-    url.searchParams.delete("ocw_resource_link_fragment")
+    url.searchParams.delete(queryKeys.uuid)
+    url.searchParams.delete(queryKeys.suffix)
     return url.toString()
   }
 
@@ -80,8 +97,8 @@ export default class ResourceLinkMarkdownSyntax extends MarkdownSyntaxPlugin {
               throw new Error("resource_link shortcode must have a uuid")
             }
             const text = escapeShortcodes(shortcode.get(1) ?? "")
-            const fragment = shortcode.get(2) ?? ""
-            const href = this.makeResourceLinkHref(uuid, fragment)
+            const suffix = shortcode.get(2) ?? ""
+            const href = this.makeResourceLinkHref(uuid, suffix)
             return `<a href="${href}">${text}</a>`
           }
         }
@@ -103,10 +120,10 @@ export default class ResourceLinkMarkdownSyntax extends MarkdownSyntaxPlugin {
             const anchor = node as HTMLAnchorElement
             const url = new URL(anchor.href)
             const search = new URLSearchParams(url.search)
-            const uuid = search.get("ocw_resource_link_uuid")
-            const hash = search.get("ocw_resource_link_fragment") ?? ""
+            const uuid = search.get(queryKeys.uuid)
+            const suffix = search.get(queryKeys.suffix) ?? ""
             if (!uuid) {
-              throw new Error("ocw_resource_link_uuid not found in URL")
+              throw new Error(`${queryKeys.suffix} not found in URL`)
             }
 
             const text = turndownService
@@ -117,7 +134,7 @@ export default class ResourceLinkMarkdownSyntax extends MarkdownSyntaxPlugin {
                */
               .replace(/{{\\</g, "{{<")
 
-            return Shortcode.resourceLink(uuid, text, hash).toHugo()
+            return Shortcode.resourceLink(uuid, text, suffix).toHugo()
           }
         }
       }
